requirements.txt
pylint
numpy
pandas
snscrape
transformers
nltk
nest_asyncio
pydantic
tweepy
python-dotenv